Add Hotel and Room model tests with comprehensive docstrings for clarity and structure
